feat(notification): refactor de notificaciones y capa service con patrones escalables
- Desacople de canales y providers (Strategy)
- Implementación de NotificationProviderRegistry (Registry)
- Refactor de channels para resolución dinámica de providers
- Mejora del TemplateBuilder usando NotificationMessage (Builder)
- Enriquecimiento de NotificationExecutionResult con metadata del provider
- Refactor de la capa service (factory, policy, helper, vo) para mejor separación de responsabilidades
- Configuración de providers por canal en application.yml

+++ btg-funds/src/main/java/com/btg/btg_funds/document/FundDocument.java
@@ -19,7 +19,7 @@ public class FundDocument {
 
     private String name;
 
-    private Integer minimumAmount;
+    private double minimumAmount;
 
     private String category;
